+++ src/@types/dto/ParametroDto.ts
@@ -10,12 +10,10 @@ export type ParametroDto = {
 };
 
 export type AtualizarParametroDto = {
-  id: number;
   tempoMedioEnvio?: number;
   tempoMedioConsumo?: number;
   tempoMedioReposicao?: number;
   quantidadeCritica?: number;
 };
 
-export type AtualizarParametroSemIdDto = Omit<AtualizarParametroDto, "id">;
 export type CriarParametroDto = Omit<ParametroDto, "id">;