+++ src/@types/repositories/IColaboradorRepository.ts
@@ -4,5 +4,5 @@ export interface IColaboradorRepository {
   save(colaborador: Colaborador): Promise<Colaborador>;
   findById(id: number): Promise<Colaborador>;
   findAll(): Promise<Colaborador[]>;
-  remove(entities0: Colaborador | Colaborador[]): Promise<Colaborador[]>;
+  remove(entities: Colaborador | Colaborador[]): Promise<Colaborador[]>;
 }