+++ src/dataMappers/colaboradorFactory.ts
@@ -1,13 +1,15 @@
-import { ColaboradorDto } from "../@types/dto/ColaboradorDto";
-import { Colaborador } from "models/ColaboradorEntity";
+import { CriarColaboradorDto } from "../@types/dto/ColaboradorDto";
+import { Colaborador } from "../models/ColaboradorEntity";
+import { enderecoFactory } from "./enderecoFactory";
 
 export const colaboradorFactory = (
-  novoColaborador: ColaboradorDto
+  novoColaborador: CriarColaboradorDto
 ): Colaborador => {
   const colaborador = new Colaborador();
   colaborador.nome = novoColaborador.nome;
   colaborador.email = novoColaborador.email;
   colaborador.telefone = novoColaborador.telefone;
-  colaborador.dataInicio = novoColaborador.dataInicio;
+  colaborador.dataInicio = new Date(novoColaborador.dataInicio);
+  colaborador.endereco = enderecoFactory(novoColaborador.endereco);
   return colaborador;
 };