+++ src/models/EnderecoEntity.ts
@@ -22,7 +22,7 @@ export class Endereco {
   complemento: string;
 
   @Column()
-  numero: string;
+  numero: number;
 
   @Column()
   bairro: string;
@@ -30,7 +30,9 @@ export class Endereco {
   @Column()
   estado: string;
 
-  @OneToOne(() => Colaborador)
+  @OneToOne(() => Colaborador, (colaborador) => colaborador.endereco, {
+    onDelete: "CASCADE",
+  })
   @JoinColumn()
   colaborador: Colaborador;
 }