+++ src/services/ColaboradorService.ts
@@ -3,11 +3,14 @@ import { IColaboradorService } from "../@types/services/IColaboradorService";
 import { Inject, Service } from "typedi";
 import { Colaborador } from "../models/ColaboradorEntity";
 import {
-  AlteraColaboradorDto,
-  ColaboradorDto,
+  AlterarColaboradorDto,
+  RetornoColaboradorCriadoDto,
+  CriarColaboradorDto,
 } from "../@types/dto/ColaboradorDto";
 import { colaboradorFactory } from "../dataMappers/colaboradorFactory";
 import { ColaboradorNaoExiste } from "../@types/errors/ColaboradorNaoExiste";
+import { RetornoEnderecoCriadoDto } from "../@types/dto/EnderecoDto";
+import { Endereco } from "../models/EnderecoEntity";
 
 @Service("ColaboradorService")
 export class ColaboradorService implements IColaboradorService {
@@ -15,27 +18,39 @@ export class ColaboradorService implements IColaboradorService {
     @Inject("ColaboradorRepository")
     private colaboradorRepository: IColaboradorRepository
   ) {}
-  async listar(): Promise<Colaborador[]> {
-    return await this.colaboradorRepository.findAll();
+  async listar(): Promise<RetornoColaboradorCriadoDto[]> {
+    const colaboradores = await this.colaboradorRepository.findAll();
+    const colaboradoresTratados = colaboradores.map((colaborador) => {
+      return this.removeIds(colaborador);
+    });
+    return colaboradoresTratados;
   }
-  async buscar(id: number): Promise<Colaborador> {
-    return await this.checaColaborador(id);
+  async buscar(colaboradorId: number): Promise<RetornoColaboradorCriadoDto> {
+    const colaborador = await this.checaColaborador(colaboradorId);
+    const colaboradorTratado = this.removeIds(colaborador);
+    return colaboradorTratado;
   }
-  async criar(colaboradorDto: ColaboradorDto): Promise<Colaborador> {
+  async criar(
+    colaboradorDto: CriarColaboradorDto
+  ): Promise<RetornoColaboradorCriadoDto> {
     const novoColaborador = colaboradorFactory(colaboradorDto);
-    return await this.colaboradorRepository.save(novoColaborador);
+    await this.colaboradorRepository.save(novoColaborador);
+    const colaboradorTratado = this.removeIds(novoColaborador);
+    return colaboradorTratado;
   }
   async atualizar(
     id: number,
-    colaboradorDtoAtualizado: AlteraColaboradorDto
-  ): Promise<void> {
+    colaboradorDtoAtualizado: AlterarColaboradorDto
+  ): Promise<RetornoColaboradorCriadoDto> {
     const colaborador = await this.checaColaborador(id);
-    const colaboradorAtualizado = {
-      ...colaborador,
-      ...colaboradorDtoAtualizado,
-    };
-    await this.colaboradorRepository.save(colaboradorAtualizado);
-    return;
+    const colaboradorAtualizado = this.atualizaColaborador(
+      colaborador,
+      colaboradorDtoAtualizado
+    );
+    const colaboradorSalvo = await this.colaboradorRepository.save(
+      colaboradorAtualizado
+    );
+    return this.removeIds(colaboradorSalvo);
   }
   async remover(id: number): Promise<void> {
     const colaboradorPraRemover = await this.checaColaborador(id);
@@ -49,5 +64,37 @@ export class ColaboradorService implements IColaboradorService {
       throw new ColaboradorNaoExiste();
     }
     return colaborador;
+  }
+
+  private removeIds(colaborador: Colaborador): RetornoColaboradorCriadoDto {
+    const { id, endereco, equipamentos, movimentacoes, ...colaboradorTratado } =
+      colaborador;
+    const novoColaborador: RetornoColaboradorCriadoDto = {
+      ...colaboradorTratado,
+      ...{ endereco: this.removeEnderecoId(endereco) },
+    };
+    return novoColaborador;
+  }
+
+  private removeEnderecoId(endereco: Endereco): RetornoEnderecoCriadoDto {
+    const { id, ...enderecoTratado } = endereco;
+    return enderecoTratado;
+  }
+  private atualizaColaborador(
+    colaborador: Colaborador,
+    colaboradorAlterado: AlterarColaboradorDto
+  ): Colaborador {
+    const { email, endereco, nome, telefone, dataInicio } = {
+      ...colaboradorAlterado,
+    };
+    if (dataInicio) {
+      colaborador.dataInicio = new Date(dataInicio);
+    }
+    colaborador.endereco = { ...colaborador.endereco, ...endereco };
+
+    const novasPropriedades = { email, nome, telefone };
+    const colaboradorAtualizado = { ...colaborador, ...novasPropriedades };
+
+    return colaboradorAtualizado;
   }
 }